fix(crosscheck): libc++abi exclusion + entity-scoped internal-namespace detection; split accounting module
Two Codex review findings, both real:

1. libc++abi mis-selected as the C++ runtime. `_cxx_runtime_lib`'s
   `startswith("libc++")` also matched `libc++abi.so.1`, so a leaked `std::__1`
   symbol was attributed to libc++abi when it preceded libc++ in the dependency
   list. Now excludes `libc++abi` (it is the separate ABI runtime, not libc++).

2. Internal-namespace detection scanned parameter types. A regex over the whole
   mangled name bucketed `foo(lib::detail::Type*)`
   (`_ZN3lib3fooEPN3lib6detail4TypeE`) as `internal_namespace` because a *parameter*
   mentioned `detail`. Replaced with `_entity_owner_is_internal`, which walks only
   the entity-name components up to the nested-name close (same entity-scoping as
   `_has_template_args`), so only the exported entity's own namespace counts.

The accounting helpers grew crosscheck.py past the 2000-line hard cap, so the pure
Itanium/MSVC mangled-name classification (constants + parsers + the `_account_*`
/ `_external_dependency_origin` / `_linked_library_names` helpers) moved to a new
sibling `buildsource/export_accounting.py`; crosscheck re-exports what it and the
tests use. No behaviour change from the move.

Accounting still sums to 100 % of exports. Adds regression tests for both fixes.
